Inline single-use Core.hs
This module has a single use and that use is more
namespace management so inline it.
